+++ src/hooks/index.ts
@@ -1,2 +1,4 @@
 export * from './useSetBackgroundColor';
 export * from './useIntersectionObserver';
+export * from './useBookCoverAnimation';
+export * from './useRibbonAnimation';
